pulling functions out of matrix API
